Refactor the mention logic into a module
